+++ R/rog-build.R
@@ -6,16 +6,14 @@
 #' This function also calls [rog_add_template_pkgdown()] and [rog_logo()].
 #'
 #' @inheritParams rog_actions_pkgdown_branch
-#' @param ... Arguments passed to [rog_logo()] and \CRANpkg{pkgdown}'s
-#'   [`pkgdown::build_site()`]. Common arguments for [rog_logo()] include
-#'   `overwrite`, `favicons`, `p_x`, `p_y` and `p_size`.
+#' @inheritDotParams rog_logo overwrite favicons p_x p_y p_size
+#' @inheritDotParams pkgdown::build_site examples preview new_process install
 #' @returns The function is called for its side effects and returns
 #'   `invisible(NULL)`.
+#' @seealso \CRANpkg{pkgdown}'s [pkgdown::build_site()].
 #' @family site
-#' @seealso \CRANpkg{pkgdown}'s [pkgdown::build_site()] and
-#'   [rog_add_template_pkgdown()].
-#' @encoding UTF-8
 #' @export
+#' @encoding UTF-8
 rog_build <- function(pkg = ".", ...) {
   # nocov start
   usethis::local_project(pkg, force = TRUE)
@@ -32,7 +30,11 @@ rog_build <- function(pkg = ".", ...) {
   do.call(rogtemplate::rog_logo, logo_dots)
   rogtemplate::rog_add_template_pkgdown()
 
-  do.call(pkgdown::build_site, c(list(pkg = "."), build_dots))
+  do.call(build_pkgdown_site, c(list(pkg = "."), build_dots))
   invisible()
   # nocov end
+}
+
+build_pkgdown_site <- function(...) {
+  pkgdown::build_site(...)
 }