+++ R/rogtemplate-package.R
@@ -1,3 +1,6 @@
+#' @seealso [rog_add_template_pkgdown()] to configure a site,
+#'   [rog_actions_pkgdown_branch()] to configure deployment and [rog_build()]
+#'   to build the site locally.
 #' @keywords internal
 #' @encoding UTF-8
 "_PACKAGE"