+++ man/roxygen/meta.R
@@ -1,7 +1,7 @@
 list(
   rd_family_title = list(
-    site = "site setup helpers",
-    assets = "package asset helpers",
+    site = "Site setup helpers",
+    assets = "Package asset helpers",
     palettes = "rOpenGov color palettes"
   )
 )